版本提升

+++ OneMinute/OneMinute/DLYRecordViewController.m
@@ -2625,7 +2625,7 @@ typedef void(^CompProgressBlcok)(CGFloat progress);
     self.timeNumber.clipsToBounds = YES;
     [_timeView addSubview:self.timeNumber];
     ////
-    NSString *partTitle = [NSString stringWithFormat:@"第%ld段",part.partNum + 1];
+    NSString *partTitle = [NSString stringWithFormat:@"第%ld段",(long)part.partNum + 1];
     NSString *timeTitle = [NSString stringWithFormat:@"%@秒",timeArr[0]];
     NSString *typeTitle;
     if (part.recordType == DLYMiniVlogRecordTypeNormal) {
